\name{HGLMLikeDeriv}
\Rdversion{1.1}
\alias{HGLMLikeDeriv}

\title{
	Display Derivatives of Log-Likelihood Functions
}

\description{Displays derivatives of fixed effects in the mean structure and dispersion parameters in the dispersion structure upon convergence.}

\usage{
HGLMLikeDeriv(x)
}

\arguments{
  \item{x}{an object inheriting from class \code{HGLM}.} 
}

\author{Marek Molas \email{m.molas@erasmusmc.nl}}



